08.11.2016
1st fully working prototype

+++ musicanal/main.py
@@ -8,37 +8,45 @@ import spotipy
 import musicdownloader
 import gmusicapi
 import auth
+import json
+import string
 
 
 
 if __name__ == '__main__':
     auth
-
+    testedtraks=dict()
     print(datetime.datetime.now())
-    dir = ''
+    dir = 'D:/lib/'
     
-    data=SPsearchplaylist.search()
+    data=auth.SPsearch()
+    '''
     with open('data.json', 'w') as f:
         json.dump(data, f, ensure_ascii=False)
     f.close()
     print(datetime.datetime.now())
-    '''
+
     with open('data.json') as f:
         data = json.load(f)
     f.close()
     print(datetime.datetime.now())
-    
-    api = musicdownloader.GMauth();
-    namelist= ['Love', 'Die']
-
-    namelist
-    for name in namelist:
-        musicdownloader.getsong(api, name)
-        sound = AudioSegment.from_mp3(dir  + name + '.mp3')
-        sound.export(dir + name + '.wav', format="wav")
-        bmpanaliz = bpmdetect.bpmdetect(dir + str(name+'.wav'))
-        os.remove(dir + name + '.wav')
-        print(datetime.datetime.now())
-
-    musicdownloader.logout(api)
     '''
+    api = musicdownloader.GMauth();
+    for namelist in data:
+        for track in namelist:
+            name=track['track']['name']
+            if track['track']['id'] not in testedtraks:
+                try:
+                    musicdownloader.getsong(api, name, dir)
+                    sound = AudioSegment.from_mp3(dir  + name + '.mp3')
+                    sound.export(dir + name + '.wav', format="wav")
+                    testedtraks[track['track']['id']]=bpmdetect.bpmdetect(dir + str(name+'.wav'))
+                    os.remove(dir + name + '.wav')
+                except:
+                    print('not found')
+            print(str(name).encode("UTF-8"))
+        print(datetime.datetime.now())
+    musicdownloader.logout(api)
+    with open('data.json', 'w') as f:
+        json.dump(testedtraks, f, ensure_ascii=False)
+    f.close()
